Fixing warnings

+++ spring-webmvc/src/test/java/org/springframework/web/servlet/tags/form/RadioButtonTagTests.java
@@ -266,7 +266,7 @@ public class RadioButtonTagTests extends AbstractFormTagTests {
 		}
 
 		public String getAsText() {
-			return "F" + (Float) getValue();
+			return "F" + getValue();
 		}
 	}
